# internal/handlers/UserHandler.go
package handlers

import (
	"encoding/json"
	"leafall/todo-service/internal/dto"
	handlersUtils "leafall/todo-service/internal/handlers/utils"
	"leafall/todo-service/internal/services"
	"leafall/todo-service/utils/exceptions"
	"leafall/todo-service/utils/query"
	"net/http"
	"strconv"

	"github.com/go-chi/chi"
)

type UserHandler struct {
	Service services.UserServiceInterface
}

func NewUserHandler(service *services.UserService) *UserHandler {
	return &UserHandler{Service: service}
}

// GetUsers godoc
// @Summary Get all users
// @Description Get all users from the database with pagination
// @Tags users
// @Accept  json
// @Produce  json
// @Param limit query int false "Limit number of users" default(10) minimum(1) maximum(100)
// @Param page query int false "Page number" default(1) minimum(1)
// @Success 200 {array} dto.UserShortResponse
// @Failure 400 {object} exceptions.ErrorResponse
// @Failure 401 {object} exceptions.ErrorResponse
// @Security BearerAuth
// @Router /v1/users [get]
func (u *UserHandler) FindAllPagination(w http.ResponseWriter, r *http.Request) {
	limit := query.GetQueryInt(r, "limit", 10)
	page := query.GetQueryInt(r, "page", 1)
	dto := dto.UserFindAllRequest{Limit: limit, Page: page}
	users, totalCount, err := u.Service.FindAllPagination(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	w.Header().Set("X-Total-Count", strconv.FormatInt(totalCount, 10))
	json.NewEncoder(w).Encode(users)
}

// FindById godoc
// @Summary Get by id
// @Description Get by id user from the database with pagination on tasks
// @Tags users
// @Accept  json
// @Produce  json
// @Param limit query int false "Limit number of tasks" default(10) minimum(1) maximum(100)
// @Param page query int false "Page number" default(1) minimum(1)
// @Param id path int true "User id" minimum(0)
// @Success 200 {object} dto.UserFullResponse
// @Failure 400 {object} exceptions.ErrorResponse
// @Failure 401 {object} exceptions.ErrorResponse
// @Security BearerAuth
// @Router /v1/users/{id} [get]
func (u *UserHandler) FindById(w http.ResponseWriter, r *http.Request) {
	limit := query.GetQueryInt(r, "limit", 10)
	page := query.GetQueryInt(r, "page", 1)
	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
    if err != nil {
        handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid user id", err.Error()), w)
        return
    }
	dto := dto.UserFindByIdRequest{Id: id, Limit: limit, Page: page}
	user, totalCount, err := u.Service.FindById(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	w.Header().Set(handlersUtils.TotalCountHeader, string(totalCount))
	json.NewEncoder(w).Encode(user)
}

// Create godoc
// @Summary Create user
// @Description Create user from the database with pagination on tasks
// @Tags users
// @Accept  json
// @Produce  json
// @Param user body dto.UserCreateRequest false "Body of request"
// @Success 200 {object} dto.UserShortResponse
// @Failure 400 {object} exceptions.ErrorResponse
// @Failure 401 {object} exceptions.ErrorResponse
// @Router /v1/users [post]
func (u *UserHandler) Create(w http.ResponseWriter, r *http.Request) {
	var dto dto.UserCreateRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}

	user, err := u.Service.Create(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(user)
}

func (u *UserHandler) Update(w http.ResponseWriter, r *http.Request) {
	var dto dto.UserUpdateRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}

	user, err := u.Service.Update(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(user)
}

func (u *UserHandler) UpdatePassword(w http.ResponseWriter, r *http.Request) {
	var dto dto.UserUpdatePasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}

	user, err := u.Service.UpdatePassword(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(user)
}

func (u *UserHandler) Delete(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
    if err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
        return
    }
	dto := &dto.UserDeleteByIdRequest{Id: id}
	result, err := u.Service.Delete(r.Context(), dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(result)
}

func (u *UserHandler) GenerateAccessToken(w http.ResponseWriter, r *http.Request) {
	var dto dto.RefreshAccessTokenRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}
	result, err := u.Service.GenerateAccessTokenFromRefreshToken(&dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(result)
}

func (u *UserHandler) Authorizate(w http.ResponseWriter, r *http.Request) {
	var dto dto.AuthorizationRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}
	result, err := u.Service.Authorizate(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(result)
}

func (u *UserHandler) Logout(w http.ResponseWriter, r *http.Request) {
	var dto dto.LogoutRequest
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		handlersUtils.HandleError(
			exceptions.NewBadRequestSimple("Invalid JSON", err.Error()), w)
		return
	}
	err := u.Service.Logout(r.Context(), &dto)
	if err != nil {
		handlersUtils.HandleError(err, w)
		return
	}
	json.NewEncoder(w).Encode(true)
}